ci: Use testcontainer for running tests
Instead of relying on services to be set up properly, use
`testcontainers` to create a container to run the test inside.

Tests can use either `TimescaleDBTestCase` or `PostgreSQLTestCase` to
select what kind of container to use.  If a Postgres container is
requested, it will be read from the environment variable
`TEST_CONTAINER_POSTGRES` and Timescale containers will be read from
`TEST_CONTAINER_TIMESCALE`.