feat: check for duplicate indexes
Add rule to check for duplicate indexes. Two indexes are duplicates if
they have different relation OID and are indexes for the same relation
and has *exactly* the same list of keys in the same order.

Note that we are not checking if an index is redundant (if one index
has a set of key columns that is a prefix of the other).

Resolves #42